requirements.txt
aiogram>=3.13.1,<4.0.0
httpx>=0.27.0,<1.0.0
beautifulsoup4>=4.12.0,<5.0.0
lxml>=5.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
tenacity>=9.0.0,<10.0.0
aiofiles>=24.1.0,<25.0.0
aiosqlite>=0.20.0,<1.0.0
# --- Dev ---
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.24.0,<1.0.0
# Примечание: на Windows lxml может потребовать
# pip install lxml --only-binary=all
